Company.java - added Vaccine methods, for creation, validation, saving and adding vaccines.
TypeVaccine.java - removed "setAdministration", added "getCode" and "getTechnology".

Vaccine.java - removed "String name" from constructor

VaccineAdministrationController.java - added vaccine methods for vaccine creation.

VaccineAdministrationUI.java - added option to choose a type of vaccine to specify vaccine administration. Code optimization with "validation" method.

Changed "dosage" from "double" to "integer" list type - US013.java, VaccineAdministration.java, VaccineAdministrationUI.java, Validate.java.

+++ src/main/java/app/domain/model/VaccineAdministration.java
@@ -9,7 +9,7 @@ public class VaccineAdministration {
 
     private List<Integer> maxAge;
 
-    private List<Double> dosage;
+    private List<Integer> dosage;
 
     private List<Integer> doses;
 
@@ -27,7 +27,7 @@ public class VaccineAdministration {
      * Creates a vaccine admnistration
      */
 
-    public VaccineAdministration(String brand, List<Integer> minAge, List<Integer> maxAge, List<Double> dosage, List<Integer> doses, List<Integer> vaccineInterval) {
+    public VaccineAdministration(String brand, List<Integer> minAge, List<Integer> maxAge, List<Integer> dosage, List<Integer> doses, List<Integer> vaccineInterval) {
         this.brand = brand;
         this.minAge = minAge;
         this.maxAge = maxAge;
@@ -45,7 +45,7 @@ public class VaccineAdministration {
 
     public List<Integer> getMaxAge() { return maxAge; }
 
-    public List<Double> getDosage() {
+    public List<Integer> getDosage() {
         return dosage;
     }
